Create bashrc
Nothing to write home about here. Just another bashrc profile but note a few things: I have created a shortcut for the ping command - in the way of a reference to another file on my host... this is very useful and I will likewise include that as well. Ping is good enough as is, a standard network utility but I have tweaked it a bit for the system administrator or network engineer hoping to be more efficient. There are a multitude of aliases and functions here as well. 

The other part of this which is useful is just my banner - and this isn't valuable from a production standpoint but if you are like me and become easily bored with vanilla applications and terminals - you might find something uniquely interesting here.