Create rsync-limit.sh
This is a basic script I wrote that attempts to calculate the anticipated growth of a file being transferred and adjust the transfer activity in correlation with the size as a it grows. This script is in the making and while it can be used in production, doing so requires that the wildcards be removed - as expansion here is not selective... 

Eventually, this will be more precise and a GUI will be built but by that point, the program will likely be written in Go, Perl or Python.... I have toyed with the idea of using Ruby but it will probably be written in Go or Python. Anyway - for now, the proof of concept is there ... More later .