Update scp-transfer
This is a script I wrote to efficiently manage file transfers - and I picked rsync as it is the file transfer method of choice when an administrator is seeking a solution that is fast, secure and reliable. In addition rsync gives administrators an answer of what to use given the deprecation of scp as a secure protocol considering Red Hat's advice to avoid it - and there is at least one CVE which speaks to that advice. I am in agreement with it.  

This script is not perfect and the wildcards I am using do not selectively sort out files by name - and for that reason, while it is great for personal use, without some tweaking it should not be used hardcore in production environments. While it is safe, I am not done perfecting it and there are some features it lacks. I will be rewriting this in likely Go or Python. You could also do something very cool by using Ruby or Perl - and given Perl's nature of handling RegEx and its ties to the Unix/Linux tradition, it would be a great language to use.