Create sed-sub
This script replaces white spaces and special characters in Linux file titles by using regex expressions. Files with white space and characters such as "!" are difficult to manipulate in the command line on Linux. I wrote this script as a way to deal with that. 

NOTE: This is a work in progress. Expect edits in the future.