Create create-user
Used to create user names and passwords - this is such a simple script - if it fails it fails - and the inverse is also true. There's no way to screw this up since the background work is handled by the internal tools to implement password creation. 

The conditional error messages are generated by the passwd and useradd commands, respectively - not the code that I included - hence, much of this code will never come into play. 

'