Create shutdown-reboot
This is a basic admin script I wrote as a method for system administrators to quickly determine whether or not a server or work station may have experienced an ungraceful shutdown. 

Much more needs to be done to pinpoint the ultimate cause - but this is a great place to start for someone who is in the dark and beginning to research the issue following an unknown point of failure.