Update stopwrite
Caution: If you do not understand this code, stop here. This is a productive script that will aid any seasoned system administrator. This is not a script for a student, someone new to UNIX/Linux, a person who does not now how the chattr command works or anyone who cannot tell apart the root file system on Linux from the user space that starts at "$HOME"