Update search-install
This is yet a third commit on this code ... I think I am getting it closer to where it needs to be and possibly finished. The last iteration of this was skipping a conditional block even in cases where the results were meaningful, despite increased user interaction which was desired. 

This version produces helpful output and does work as intended.