Update search-install
Regular update. This time I have checked for more unexpected errors - and this is always difficult when using command line scripts in general as it is hard to separated output desired from the code and redundant output generated by the command line tools which is independent from the behavior I aim for.