Update mountedtotals
Please see the comments. This was a tough script to write but it is helpful and I hope you find help with it, whomever you are reading this. I encourage anyone and everyone to tinker with this and modify it. If you want to learn more about the Linux file system, this is a good way to start. You'll learn as much seeing the limitations of this as you will in the utility of it.