Create ftp-fileselect
I'm proud of this. Enjoy if you are reading it. This can be deployed in production - but... I will not be liable for any issues if the person reading this does not properly evaluate and review the code. 

You should test anything before deploying it and, on that note, this is an unfinished script. It works well enough to use but I have not finished fully debugging it. I will try to update this sooner rather than later if there is a need for that.